add reused getsharedpoints code, do some cleanup based on Xcode warnings

+++ MobileLighting_Mac/calibration/calibration/calib_utils.cpp
@@ -141,8 +141,8 @@ vector<vector<int>> extractIds( const FileNode& array ) {
     return output;
 }
 Mat extractMatrix( const FileNode& array ) {
-    int rows = array.size();
-    int cols = array[0].size();
+    size_t rows = array.size();
+    size_t cols = array[0].size();
     Mat m = Mat( rows, cols, CV_32F );
     for(int i = 0; i < rows; i++) {
         for(int j = 0; j < cols; j++) {
